Add the configuration constant WC1C_PRESERVE_PRODUCT_VARIATIONS
git-svn-id: https://plugins.svn.wordpress.org/woocommerce-and-1centerprise-data-exchange/trunk@1338250 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ exchange/offers.php
@@ -2,6 +2,7 @@
 if (!defined('ABSPATH')) exit;
 
 if (!defined('WC1C_PRICE_TYPE')) define('WC1C_PRICE_TYPE', null);
+if (!defined('WC1C_PRESERVE_PRODUCT_VARIATIONS')) define('WC1C_PRESERVE_PRODUCT_VARIATIONS', false);
 
 function wc1c_offers_start_element_handler($is_full, $names, $depth, $name, $attrs) {
   global $wc1c_price_types, $wc1c_offer, $wc1c_price;
@@ -340,8 +341,10 @@ function wc1c_replace_suboffers($is_full, $suboffers, $are_products = false) {
     }
   }
 
-  $deleted_product_variation_ids = array_diff($current_product_variation_ids, $product_variation_ids);
-  foreach ($deleted_product_variation_ids as $deleted_product_variation_id) {
-    wp_delete_post($deleted_product_variation_id, true);
+  if (!WC1C_PRESERVE_PRODUCT_VARIATIONS) {
+    $deleted_product_variation_ids = array_diff($current_product_variation_ids, $product_variation_ids);
+    foreach ($deleted_product_variation_ids as $deleted_product_variation_id) {
+      wp_delete_post($deleted_product_variation_id, true);
+    }
   }
 }
